backend/requirements.txt
django
python-dotenv
djangorestframework
drf-spectacular
pillow
django-cors-headers
channels
uvicorn[standard]